solaris systems 64 bits build fix proposal.

+++ libhfcommon/util.c
@@ -156,7 +156,7 @@ bool util_PinThreadToCPUs(uint32_t threadno, uint32_t cpucnt) {
     for (uint32_t i = 0; i < cpucnt; i++) {
         if (pset_assign(set, ((start_cpu + i) % num_cpus), NULL) != 0) {
             PLOG_W("pset_assign(%" PRIu32 "), failed", i);
-            pset_destroy(set);
+            pset_destroy(set); // TODO: defer mechanism not yet supported
             return false;
         }
     }
@@ -164,7 +164,7 @@ bool util_PinThreadToCPUs(uint32_t threadno, uint32_t cpucnt) {
     p = getpid();
 
     if (pset_bind(set, P_PID, p, NULL) != 0) {
-        PLOG_W("pset_bind(%ld) failed", p);
+        PLOG_W("pset_bind(%ld) failed", (long)p);
         ret = false;
     }
